+++ src/lib.rs
@@ -34,6 +34,9 @@ pub mod jetstream;
 #[cfg(feature = "jetstream")]
 pub mod object_store;
 
+#[cfg(feature = "jetstream")]
+pub mod schedule;
+
 #[cfg(feature = "kv")]
 pub mod kv;